refactor consensuser tests

+++ consensuser/configuration.go
@@ -8,9 +8,9 @@ type Configuration struct {
 	Info cont.Set
 }
 
-func NewMasterlessConfiguration(numberNodes uint32) Configuration {
+func NewMasterlessConfiguration(numberNodes int) Configuration {
 	conf := new(Configuration)
-	conf.Info = cont.NewSet(numberNodes)
+	conf.Info = cont.NewSet(uint32(numberNodes))
 	return *conf
 }
